chore: add secure route for user to get game history
